code review & light refactorization
