+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.13)
-##############################################################################################################
+cmake_policy(VERSION 3.13)
 
 project(fty-outage
   VERSION 1.0.0
@@ -12,6 +12,8 @@ find_package(fty-cmake PATHS ${CMAKE_BINARY_DIR}/fty-cmake)
 
 etn_target(static ${PROJECT_NAME}-lib
     SOURCES
+        expiration.h
+        expiration.cc
         data.h
         data.cc
         fty-outage-server.h
@@ -64,6 +66,7 @@ etn_configure_file(
 etn_test_target(${PROJECT_NAME}-lib
     SOURCES
         test/main.cpp
+        test/expiration.cpp
         test/data.cpp
         test/outage.cpp
     PREPROCESSOR
