+++ resources/fty-outage.conf.in
@@ -1,3 +1,3 @@
-# create runtime directories for agent fty-outage
+# create runtime directories for agent @PROJECT_NAME@
 d /var/lib/fty/fty-outage 0755 bios root
 x /var/lib/fty/fty-outage/*
